Move buttons to button subpackage. Fix bug (change getInteger to parseInt)
