+++ src/main/java/gui/hashmap/panel/button/PutButton.java
@@ -1,6 +1,7 @@
-package gui.hashmap.panel.button.listener;
+package gui.hashmap.panel.button;
 
 import gui.hashmap.panel.HashMapPanel;
+import gui.hashmap.panel.button.listener.PutListener;
 import gui.hashmap.panel.image.ImageProducer;
 
 import javax.swing.*;
